+++ client/my-sites/plugins/controller.js
@@ -18,9 +18,6 @@ import PluginBrowser from './plugins-browser';
 import PluginUpload from './plugin-upload';
 import { getSelectedSite } from 'calypso/state/ui/selectors';
 import getSelectedOrAllSitesWithPlugins from 'calypso/state/selectors/get-selected-or-all-sites-with-plugins';
-import AsyncLoad from 'calypso/components/async-load';
-import MarketplacePluginDetails from 'calypso/my-sites/plugins/marketplace/marketplace-plugin-details';
-
 /**
  * Module variables
  */
@@ -47,31 +44,6 @@ function renderSinglePlugin( context, siteUrl ) {
 		pluginSlug,
 		siteUrl,
 	} );
-}
-
-export function renderMarketplacePlugin( context, next ) {
-	const siteUrl = getSiteFragment( context.path );
-
-	const pluginSlug = decodeURIComponent( context.params.plugin );
-
-	let prevPath;
-	if ( lastPluginsListVisited ) {
-		prevPath = lastPluginsListVisited;
-	} else if ( context.prevPath ) {
-		prevPath = sectionify( context.prevPath );
-	}
-
-	context.primary = (
-		<MarketplacePluginDetails
-			path={ context.path }
-			prevQuerystring={ lastPluginsQuerystring }
-			prevPath={ prevPath }
-			marketplacePluginSlug={ pluginSlug }
-			siteUrl={ siteUrl }
-		/>
-	);
-
-	next();
 }
 
 function getPathWithoutSiteSlug( context, site ) {
@@ -226,19 +198,5 @@ export function scrollTopIfNoHash( context, next ) {
 	if ( typeof window !== 'undefined' && ! window.location.hash ) {
 		window.scrollTo( 0, 0 );
 	}
-	next();
-}
-
-export function renderDomainsPage( context, next ) {
-	context.primary = (
-		<AsyncLoad require="calypso/my-sites/plugins/marketplace/marketplace-domain-upsell" />
-	);
-	next();
-}
-
-export function renderPluginsSetupStatusPage( context, next ) {
-	context.primary = (
-		<AsyncLoad require="calypso/my-sites/plugins/marketplace/marketplace-plugin-setup-status" />
-	);
 	next();
 }
