+++ client/sections.js
@@ -109,6 +109,13 @@ const sections = [
 		group: 'sites',
 	},
 	{
+		// TODO: Move this out to its own root folder
+		name: 'marketplace',
+		paths: [ '/marketplace' ],
+		module: 'calypso/my-sites/plugins/marketplace',
+		group: 'sites',
+	},
+	{
 		name: 'pages',
 		paths: [ '/pages' ],
 		module: 'calypso/my-sites/pages',
